+++ tests/Integration/MSFT_SqlRSSetup.Integration.Tests.ps1
@@ -12,7 +12,7 @@ $script:dscResourceName = "MSFT_$($script:dscResourceFriendlyName)"
 
 try
 {
-    Import-Module -Name DscResource.Test -Force
+    Import-Module -Name DscResource.Test -Force -ErrorAction 'Stop'
 }
 catch [System.IO.FileNotFoundException]
 {