+++ tests/Unit/MSFT_SqlScript.Tests.ps1
@@ -26,7 +26,7 @@ function Invoke-TestSetup
 {
     try
     {
-        Import-Module -Name DscResource.Test -Force
+        Import-Module -Name DscResource.Test -Force -ErrorAction 'Stop'
     }
     catch [System.IO.FileNotFoundException]
     {