+++ tests/Unit/MSFT_SqlServerConfiguration.Tests.ps1
@@ -20,7 +20,7 @@ $script:dscResourceName = 'MSFT_SqlServerConfiguration'
 
 try
 {
-    Import-Module -Name DscResource.Test -Force
+    Import-Module -Name DscResource.Test -Force -ErrorAction 'Stop'
 }
 catch [System.IO.FileNotFoundException]
 {