+++ tests/Unit/MSFT_SqlServerReplication.Tests.ps1
@@ -24,7 +24,7 @@ $script:dscResourceName = 'MSFT_SqlServerReplication'
 
 try
 {
-    Import-Module -Name DscResource.Test -Force
+    Import-Module -Name DscResource.Test -Force -ErrorAction 'Stop'
 }
 catch [System.IO.FileNotFoundException]
 {