+++ tests/Unit/MSFT_SqlServerRole.Tests.ps1
@@ -22,7 +22,7 @@ function Invoke-TestSetup
 {
     try
     {
-        Import-Module -Name DscResource.Test -Force
+        Import-Module -Name DscResource.Test -Force -ErrorAction 'Stop'
     }
     catch [System.IO.FileNotFoundException]
     {